Order of schema Publish-Fetch inverted. SSL certs added. Local Schema path fixed
The adapter now tries to publish local schemas before fetching from broker. This fixes the problem of a topic not registered in broker but locally and required by the adapter.
A new set of SSL certificates have been added based on the sample-ssl-certificates from local+security docker instance. This has been tested to work on default local+security.
The default path to local schema now works on both linux and windows.